'Fix' token cookies being set to undefined on the server. Remove undefined cookies in the browser. Fix CartItem serialization failing due to missing price

+++ Angular/src/app/admin/admin.service.ts
@@ -23,7 +23,7 @@ export class AdminService {
 
 	public getAllUsers(page?: number, size?: number): Observable<User[]> {
 		const accessToken = this.cookie.get('accessToken');
-		if (accessToken) {
+		if (accessToken && accessToken !== 'undefined') {
 			const headers = new HttpHeaders().append('Authorization', 'Bearer ' + accessToken);
 			let params = new HttpParams();
 			if (page && size) {
@@ -37,7 +37,7 @@ export class AdminService {
 
 	public getUser(id: string): Observable<User> {
 		const accessToken = this.cookie.get('accessToken');
-		if (accessToken) {
+		if (accessToken && accessToken !== 'undefined') {
 			const headers = new HttpHeaders().append('Authorization', 'Bearer ' + accessToken);
 			return this.http.get<User>(this.adminBase + 'user/' + id, { headers });
 		} else {
@@ -47,7 +47,7 @@ export class AdminService {
 
 	public getUserCount(): Observable<number> {
 		const accessToken = this.cookie.get('accessToken');
-		if (accessToken) {
+		if (accessToken && accessToken !== 'undefined') {
 			const headers = new HttpHeaders().append('Authorization', 'Bearer ' + accessToken);
 			return this.http.get<number>(this.adminBase + 'users/count', { headers });
 		} else {
@@ -57,7 +57,7 @@ export class AdminService {
 
 	public getAllPages(page?: number, size?: number): Observable<Page[]> {
 		const accessToken = this.cookie.get('accessToken');
-		if (accessToken) {
+		if (accessToken && accessToken !== 'undefined') {
 			const headers = new HttpHeaders().append('Authorization', 'Bearer ' + accessToken);
 			let params = new HttpParams();
 			if (page && size) {
@@ -77,7 +77,7 @@ export class AdminService {
 
 	public getPage(id: string): Observable<Page> {
 		const accessToken = this.cookie.get('accessToken');
-		if (accessToken) {
+		if (accessToken && accessToken !== 'undefined') {
 			const headers = new HttpHeaders().append('Authorization', 'Bearer ' + accessToken);
 			return this.http.get<Page>(this.adminBase + 'page/' + id, { headers }).pipe(map(page => {
 				page.created = new Date(page.created!);
@@ -91,7 +91,7 @@ export class AdminService {
 
 	public getPageCount(): Observable<number> {
 		const accessToken = this.cookie.get('accessToken');
-		if (accessToken) {
+		if (accessToken && accessToken !== 'undefined') {
 			const headers = new HttpHeaders().append('Authorization', 'Bearer ' + accessToken);
 			return this.http.get<number>(this.adminBase + 'pages/count', { headers });
 		} else {
@@ -101,7 +101,7 @@ export class AdminService {
 
 	public submitPage(page: Page): Observable<Page> {
 		const accessToken = this.cookie.get('accessToken');
-		if (accessToken) {
+		if (accessToken && accessToken !== 'undefined') {
 			const headers = new HttpHeaders().append('Authorization', 'Bearer ' + accessToken);
 			return this.http.post<Page>(this.adminBase + 'pages', page, { headers });
 		} else {
@@ -111,7 +111,7 @@ export class AdminService {
 
 	public editPage(page: Page): Observable<string> {
 		const accessToken = this.cookie.get('accessToken');
-		if (accessToken) {
+		if (accessToken && accessToken !== 'undefined') {
 			const headers = new HttpHeaders().append('Authorization', 'Bearer ' + accessToken);
 			return this.http.put<string>(this.adminBase + 'page/' + page.id, page, { headers });
 		} else {
@@ -121,7 +121,7 @@ export class AdminService {
 
 	public deletePage(id: string): Observable<string> {
 		const accessToken = this.cookie.get('accessToken');
-		if (accessToken) {
+		if (accessToken && accessToken !== 'undefined') {
 			const headers = new HttpHeaders().append('Authorization', 'Bearer ' + accessToken);
 			return this.http.delete<string>(this.adminBase + 'page/' + id, { headers });
 		} else {
@@ -131,7 +131,7 @@ export class AdminService {
 
 	public checkPageSlugTaken(slug: string): Observable<string> {
 		const accessToken = this.cookie.get('accessToken');
-		if (accessToken) {
+		if (accessToken && accessToken !== 'undefined') {
 			const headers = new HttpHeaders().append('Authorization', 'Bearer ' + accessToken).append('Accept', 'application/json');
 			const params = new HttpParams().append('slug', slug)
 			return this.http.get<string>(this.adminBase + 'pages/slugTaken', { headers, params });
@@ -142,7 +142,7 @@ export class AdminService {
 
 	public getAllProducts(page?: number, size?: number): Observable<Product[]> {
 		const accessToken = this.cookie.get('accessToken');
-		if (accessToken) {
+		if (accessToken && accessToken !== 'undefined') {
 			const headers = new HttpHeaders().append('Authorization', 'Bearer ' + accessToken);
 			let params = new HttpParams();
 			if (page && size) {
@@ -162,7 +162,7 @@ export class AdminService {
 
 	public getProduct(id: string): Observable<Product> {
 		const accessToken = this.cookie.get('accessToken');
-		if (accessToken) {
+		if (accessToken && accessToken !== 'undefined') {
 			const headers = new HttpHeaders().append('Authorization', 'Bearer ' + accessToken);
 			return this.http.get<Product>(this.adminBase + 'product/' + id, { headers }).pipe(map(page => {
 				page.created = new Date(page.created!);
@@ -176,7 +176,7 @@ export class AdminService {
 
 	public editProduct(product: Product): Observable<string> {
 		const accessToken = this.cookie.get('accessToken');
-		if (accessToken) {
+		if (accessToken && accessToken !== 'undefined') {
 			const headers = new HttpHeaders().append('Authorization', 'Bearer ' + accessToken);
 			return this.http.put<string>(this.adminBase + 'product/' + product.id, product, { headers });
 		} else {
@@ -186,7 +186,7 @@ export class AdminService {
 
 	public getProductCount(): Observable<number> {
 		const accessToken = this.cookie.get('accessToken');
-		if (accessToken) {
+		if (accessToken && accessToken !== 'undefined') {
 			const headers = new HttpHeaders().append('Authorization', 'Bearer ' + accessToken);
 			return this.http.get<number>(this.adminBase + 'products/count', { headers });
 		} else {
@@ -196,7 +196,7 @@ export class AdminService {
 
 	public submitProduct(product: Product): Observable<Product> {
 		const accessToken = this.cookie.get('accessToken');
-		if (accessToken) {
+		if (accessToken && accessToken !== 'undefined') {
 			const headers = new HttpHeaders().append('Authorization', 'Bearer ' + accessToken);
 			return this.http.post<Product>(this.adminBase + 'products', product, { headers });
 		} else {
@@ -206,7 +206,7 @@ export class AdminService {
 
 	public deleteProduct(id: string): Observable<string> {
 		const accessToken = this.cookie.get('accessToken');
-		if (accessToken) {
+		if (accessToken && accessToken !== 'undefined') {
 			const headers = new HttpHeaders().append('Authorization', 'Bearer ' + accessToken);
 			return this.http.delete<string>(this.adminBase + 'product/' + id, { headers });
 		} else {
@@ -216,7 +216,7 @@ export class AdminService {
 
 	public checkProductSlugTaken(slug: string): Observable<string> {
 		const accessToken = this.cookie.get('accessToken');
-		if (accessToken) {
+		if (accessToken && accessToken !== 'undefined') {
 			const headers = new HttpHeaders().append('Authorization', 'Bearer ' + accessToken).append('Accept', 'application/json');
 			const params = new HttpParams().append('slug', slug)
 			return this.http.get<string>(this.adminBase + 'products/slugTaken', { headers, params });
@@ -227,7 +227,7 @@ export class AdminService {
 
 	public getAllCoupons(page?: number, size?: number): Observable<Coupon[]> {
 		const accessToken = this.cookie.get('accessToken');
-		if (accessToken) {
+		if (accessToken && accessToken !== 'undefined') {
 			const headers = new HttpHeaders().append('Authorization', 'Bearer ' + accessToken);
 			let params = new HttpParams();
 			if (page && size) {
@@ -247,7 +247,7 @@ export class AdminService {
 
 	public getCoupon(id: string): Observable<Coupon> {
 		const accessToken = this.cookie.get('accessToken');
-		if (accessToken) {
+		if (accessToken && accessToken !== 'undefined') {
 			const headers = new HttpHeaders().append('Authorization', 'Bearer ' + accessToken);
 			return this.http.get<Coupon>(this.adminBase + 'coupon/' + id, { headers }).pipe(map(coupon => {
 				coupon.created = new Date(coupon.created!);
@@ -261,7 +261,7 @@ export class AdminService {
 
 	public getCouponCount(): Observable<number> {
 		const accessToken = this.cookie.get('accessToken');
-		if (accessToken) {
+		if (accessToken && accessToken !== 'undefined') {
 			const headers = new HttpHeaders().append('Authorization', 'Bearer ' + accessToken);
 			return this.http.get<number>(this.adminBase + 'coupons/count', { headers });
 		} else {
@@ -271,7 +271,7 @@ export class AdminService {
 
 	public submitCoupon(coupon: Coupon): Observable<Coupon> {
 		const accessToken = this.cookie.get('accessToken');
-		if (accessToken) {
+		if (accessToken && accessToken !== 'undefined') {
 			const headers = new HttpHeaders().append('Authorization', 'Bearer ' + accessToken);
 			return this.http.post<Coupon>(this.adminBase + 'coupons', coupon, { headers });
 		} else {
@@ -281,7 +281,7 @@ export class AdminService {
 
 	public editCoupon(coupon: Coupon): Observable<string> {
 		const accessToken = this.cookie.get('accessToken');
-		if (accessToken) {
+		if (accessToken && accessToken !== 'undefined') {
 			const headers = new HttpHeaders().append('Authorization', 'Bearer ' + accessToken);
 			return this.http.put<string>(this.adminBase + 'coupon/' + coupon.id, coupon, { headers });
 		} else {
@@ -291,7 +291,7 @@ export class AdminService {
 
 	public deleteCoupon(id: string): Observable<string> {
 		const accessToken = this.cookie.get('accessToken');
-		if (accessToken) {
+		if (accessToken && accessToken !== 'undefined') {
 			const headers = new HttpHeaders().append('Authorization', 'Bearer ' + accessToken);
 			return this.http.delete<string>(this.adminBase + 'coupon/' + id, { headers });
 		} else {
@@ -301,7 +301,7 @@ export class AdminService {
 
 	public checkCouponSlugTaken(slug: string): Observable<string> {
 		const accessToken = this.cookie.get('accessToken');
-		if (accessToken) {
+		if (accessToken && accessToken !== 'undefined') {
 			const headers = new HttpHeaders().append('Authorization', 'Bearer ' + accessToken).append('Accept', 'application/json');
 			const params = new HttpParams().append('slug', slug)
 			return this.http.get<string>(this.adminBase + 'coupons/slugTaken', { headers, params });
@@ -312,7 +312,7 @@ export class AdminService {
 
 	public getAllOrders(page?: number, size?: number): Observable<Order[]> {
 		const accessToken = this.cookie.get('accessToken');
-		if (accessToken) {
+		if (accessToken && accessToken !== 'undefined') {
 			const headers = new HttpHeaders().append('Authorization', 'Bearer ' + accessToken);
 			let params = new HttpParams();
 			if (page && size) {
@@ -326,7 +326,7 @@ export class AdminService {
 
 	public getOrder(id: string): Observable<Order> {
 		const accessToken = this.cookie.get('accessToken');
-		if (accessToken) {
+		if (accessToken && accessToken !== 'undefined') {
 			const headers = new HttpHeaders().append('Authorization', 'Bearer ' + accessToken);
 			return this.http.get<Order>(this.adminBase + 'orders/' + id, { headers });
 		} else {
@@ -336,7 +336,7 @@ export class AdminService {
 
 	public getOrderCount(): Observable<number> {
 		const accessToken = this.cookie.get('accessToken');
-		if (accessToken) {
+		if (accessToken && accessToken !== 'undefined') {
 			const headers = new HttpHeaders().append('Authorization', 'Bearer ' + accessToken);
 			return this.http.get<number>(this.adminBase + 'orders/count', { headers });
 		} else {
@@ -346,7 +346,7 @@ export class AdminService {
 
 	public deleteOrder(id: string): Observable<string> {
 		const accessToken = this.cookie.get('accessToken');
-		if (accessToken) {
+		if (accessToken && accessToken !== 'undefined') {
 			const headers = new HttpHeaders().append('Authorization', 'Bearer ' + accessToken);
 			return this.http.delete<string>(this.adminBase + 'order/' + id, { headers });
 		} else {
@@ -356,7 +356,7 @@ export class AdminService {
 
 	public getAllShippingZones(page?: number, size?: number): Observable<ShippingZone[]> {
 		const accessToken = this.cookie.get('accessToken');
-		if (accessToken) {
+		if (accessToken && accessToken !== 'undefined') {
 			const headers = new HttpHeaders().append('Authorization', 'Bearer ' + accessToken);
 			let params = new HttpParams();
 			if (page && size) {
@@ -370,7 +370,7 @@ export class AdminService {
 
 	public getShippingZone(id: string): Observable<ShippingZone> {
 		const accessToken = this.cookie.get('accessToken');
-		if (accessToken) {
+		if (accessToken && accessToken !== 'undefined') {
 			const headers = new HttpHeaders().append('Authorization', 'Bearer ' + accessToken);
 			return this.http.get<ShippingZone>(this.adminBase + 'usShippingZone/' + id, { headers });
 		} else {
@@ -380,7 +380,7 @@ export class AdminService {
 
 	public getShippingZoneCount(): Observable<number> {
 		const accessToken = this.cookie.get('accessToken');
-		if (accessToken) {
+		if (accessToken && accessToken !== 'undefined') {
 			const headers = new HttpHeaders().append('Authorization', 'Bearer ' + accessToken);
 			return this.http.get<number>(this.adminBase + 'usShippingZones/count', { headers });
 		} else {
@@ -390,7 +390,7 @@ export class AdminService {
 
 	public submitShippingZone(shippingZone: ShippingZone): Observable<ShippingZone> {
 		const accessToken = this.cookie.get('accessToken');
-		if (accessToken) {
+		if (accessToken && accessToken !== 'undefined') {
 			const headers = new HttpHeaders().append('Authorization', 'Bearer ' + accessToken);
 			return this.http.post<ShippingZone>(this.adminBase + 'usShippingZones', shippingZone, { headers });
 		} else {
@@ -400,7 +400,7 @@ export class AdminService {
 
 	public editShippingZone(shippingZone: ShippingZone): Observable<string> {
 		const accessToken = this.cookie.get('accessToken');
-		if (accessToken) {
+		if (accessToken && accessToken !== 'undefined') {
 			const headers = new HttpHeaders().append('Authorization', 'Bearer ' + accessToken);
 			return this.http.put<string>(this.adminBase + 'usShippingZone/' + shippingZone.id, shippingZone, { headers });
 		} else {
@@ -410,7 +410,7 @@ export class AdminService {
 
 	public deleteShippingZone(id: string): Observable<string> {
 		const accessToken = this.cookie.get('accessToken');
-		if (accessToken) {
+		if (accessToken && accessToken !== 'undefined') {
 			const headers = new HttpHeaders().append('Authorization', 'Bearer ' + accessToken);
 			return this.http.delete<string>(this.adminBase + 'usShippingZone/' + id, { headers });
 		} else {
@@ -420,7 +420,7 @@ export class AdminService {
 
 	public getMenuItems(): Observable<MenuItem[]> {
 		const accessToken = this.cookie.get('accessToken');
-		if (accessToken) {
+		if (accessToken && accessToken !== 'undefined') {
 			const headers = new HttpHeaders().append('Authorization', 'Bearer ' + accessToken);
 			return this.http.get<MenuItem[]>(this.adminBase + 'menuItems', { headers });
 		} else {
@@ -430,7 +430,7 @@ export class AdminService {
 
 	public saveMenuItems(menuItems: MenuItem[]): Observable<MenuItem[]> {
 		const accessToken = this.cookie.get('accessToken');
-		if (accessToken) {
+		if (accessToken && accessToken !== 'undefined') {
 			const headers = new HttpHeaders().append('Authorization', 'Bearer ' + accessToken);
 			return this.http.post<MenuItem[]>(this.adminBase + 'menuItems', menuItems, { headers });
 		} else {
